+++ core/types/chat_messages.go
@@ -3,7 +3,7 @@ package types
 type Chat struct {
 	UserName  string `xml:"username,attr" json:"user_name"`
 	Message   string `xml:"message,attr" json:"message"`
-	Timestamp int64  `xml:"timestamp,attr" json:"timestamp"`
+	Timestamp int    `xml:"timestamp,attr" json:"timestamp"`
 }
 
 type ChatMessages struct {
