+++ core/types/scan_status.go
@@ -2,8 +2,8 @@ package types
 
 type ScanStatus struct {
 	Scanning      bool   `xml:"scanning,attr" json:"scanning"`
-	Count         int64  `xml:"count,attr" json:"count"`
-	FolderCount   int64  `xml:"folderCount,attr" json:"folderCount"`
+	Count         int    `xml:"count,attr" json:"count"`
+	FolderCount   int    `xml:"folderCount,attr" json:"folderCount"`
 	StartedDate   string `xml:"startedDate,attr" json:"startedDate"`
 	Type          string `xml:"type,attr" json:"type"`
 	CompletedDate string `xml:"completedDate,attr" json:"completedDate"`
